refactor: separate the new metric file implementation to follow the same existing patterns;
