+++ metric/requests_inflight.go
@@ -11,10 +11,6 @@ const (
 	metricNameRequestInFlight = "requests_inflight"
 	metricUnitRequestInFlight = "{count}"
 	metricDescRequestInFlight = "Measures the number of requests currently being processed by the server."
-
-	metricNameServerActiveRequests = "http.server.active_requests"
-	metricUnitServerActiveRequests = "{request}"
-	metricDescServerActiveRequests = "Number of active HTTP server requests."
 )
 
 // [RequestInFlight] is a metrics recorder for recording the number of requests in flight.
@@ -44,29 +40,6 @@ func NewRequestInFlight(cfg BaseConfig) func(next http.Handler) http.Handler {
 
 			// decrease the number of requests in flight
 			counter.Add(r.Context(), -1, attrs)
-		})
-	}
-}
-
-// NewServerActiveRequests records the number of active HTTP server requests.
-func NewServerActiveRequests(cfg BaseConfig) func(next http.Handler) http.Handler {
-	counter, err := cfg.Meter.Int64UpDownCounter(
-		metricNameServerActiveRequests,
-		otelmetric.WithDescription(metricDescServerActiveRequests),
-		otelmetric.WithUnit(metricUnitServerActiveRequests),
-	)
-	if err != nil {
-		panic(fmt.Sprintf("unable to create %s counter: %v", metricNameServerActiveRequests, err))
-	}
-
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			attrs := otelmetric.WithAttributes(cfg.AttributesFunc(r)...)
-
-			counter.Add(r.Context(), 1, attrs)
-			defer counter.Add(r.Context(), -1, attrs)
-
-			next.ServeHTTP(w, r)
 		})
 	}
 }
