+++ metric/response_size.go
@@ -11,10 +11,6 @@ const (
 	metricNameResponseSizeBytes = "response_size_bytes"
 	metricUnitResponseSizeBytes = "By"
 	metricDescResponseSizeBytes = "Measures the size of the response in bytes."
-
-	metricNameServerResponseBodySize = "http.server.response.body.size"
-	metricUnitServerResponseBodySize = "By"
-	metricDescServerResponseBodySize = "Size of HTTP server response bodies."
 )
 
 // Deprecated: use NewServerResponseBodySize instead.
@@ -42,35 +38,6 @@ func NewResponseSizeBytes(cfg BaseConfig) func(next http.Handler) http.Handler {
 			histogram.Record(
 				r.Context(),
 				int64(rrw.writtenBytes),
-				otelmetric.WithAttributes(
-					cfg.AttributesFunc(r)...,
-				),
-			)
-		})
-	}
-}
-
-// NewServerResponseBodySize records the size of HTTP server response bodies in bytes.
-func NewServerResponseBodySize(cfg BaseConfig) func(next http.Handler) http.Handler {
-	histogram, err := cfg.Meter.Int64Histogram(
-		metricNameServerResponseBodySize,
-		otelmetric.WithDescription(metricDescServerResponseBodySize),
-		otelmetric.WithUnit(metricUnitServerResponseBodySize),
-	)
-	if err != nil {
-		panic(fmt.Sprintf("unable to create %s histogram: %v", metricNameServerResponseBodySize, err))
-	}
-
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			rrw := getRRW(w)
-			defer putRRW(rrw)
-
-			next.ServeHTTP(rrw.writer, r)
-
-			histogram.Record(
-				r.Context(),
-				rrw.writtenBytes,
 				otelmetric.WithAttributes(
 					cfg.AttributesFunc(r)...,
 				),
