+++ metric/server_active_requests_test.go
@@ -1,0 +1,49 @@
+package metric_test
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/go-chi/chi/v5"
+	"github.com/riandyrn/otelchi/metric"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/attribute"
+	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata"
+)
+
+func TestServerActiveRequests(t *testing.T) {
+	reader := sdkmetric.NewManualReader()
+	provider := sdkmetric.NewMeterProvider(sdkmetric.WithReader(reader))
+	baseCfg := metric.NewBaseConfig(
+		"test-server",
+		metric.WithMeterProvider(provider),
+		metric.WithAttributesFunc(testAttributes),
+	)
+
+	router := chi.NewRouter()
+	router.Use(metric.NewServerActiveRequests(baseCfg))
+	router.Get("/test", func(w http.ResponseWriter, r *http.Request) {
+		m := findMetric(t, collectMetrics(t, reader), "http.server.active_requests")
+		assert.Equal(t, "{request}", m.Unit)
+		assert.Equal(t, "Number of active HTTP server requests.", m.Description)
+
+		sum, ok := m.Data.(metricdata.Sum[int64])
+		require.True(t, ok)
+		require.Len(t, sum.DataPoints, 1)
+		assert.Equal(t, int64(1), sum.DataPoints[0].Value)
+		assertHasAttribute(t, sum.DataPoints[0].Attributes, attribute.String("test.attr", "value"))
+
+		w.WriteHeader(http.StatusOK)
+	})
+
+	router.ServeHTTP(httptest.NewRecorder(), httptest.NewRequest(http.MethodGet, "/test", nil))
+
+	m := findMetric(t, collectMetrics(t, reader), "http.server.active_requests")
+	sum, ok := m.Data.(metricdata.Sum[int64])
+	require.True(t, ok)
+	require.Len(t, sum.DataPoints, 1)
+	assert.Equal(t, int64(0), sum.DataPoints[0].Value)
+}
