+++ metric/server_test.go
@@ -1,0 +1,50 @@
+package metric_test
+
+import (
+	"context"
+	"net/http"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/attribute"
+	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata"
+)
+
+func testAttributes(*http.Request) []attribute.KeyValue {
+	return []attribute.KeyValue{
+		attribute.String("test.attr", "value"),
+	}
+}
+
+func collectMetrics(t *testing.T, reader *sdkmetric.ManualReader) metricdata.ResourceMetrics {
+	t.Helper()
+
+	var rm metricdata.ResourceMetrics
+	err := reader.Collect(context.Background(), &rm)
+	require.NoError(t, err)
+	return rm
+}
+
+func findMetric(t *testing.T, rm metricdata.ResourceMetrics, name string) metricdata.Metrics {
+	t.Helper()
+
+	for _, sm := range rm.ScopeMetrics {
+		for _, m := range sm.Metrics {
+			if m.Name == name {
+				return m
+			}
+		}
+	}
+	t.Fatalf("metric %q not found", name)
+	return metricdata.Metrics{}
+}
+
+func assertHasAttribute(t *testing.T, attrs attribute.Set, want attribute.KeyValue) {
+	t.Helper()
+
+	got, ok := attrs.Value(want.Key)
+	require.True(t, ok)
+	assert.Equal(t, want.Value.AsString(), got.AsString())
+}
